requirements.txt
anthropic>=0.50.0
openai>=1.60.0
google-generativeai>=0.8.0
pyyaml>=6.0
python-dotenv>=1.0
tiktoken>=0.8.0
pandas>=2.0
